+++ src/marketplaces/X2Y2Mappings.ts
@@ -28,7 +28,7 @@ export function handleEvInventory(event: EvInventory): void {
 
       if (currencyEntity) {
         //Gather the decimals used in the currency transacted in
-        let amountDecimals = constants.BIGINT_TEN.pow(currencyEntity.decimals)
+        let amountDecimals = constants.BIGINT_TEN.pow(currencyEntity.decimals as u8)
 
         //4. Assign currency address, amount, txId and platform to sale entity
         let saleEntity = new sale(event.block.number.toString() + '-' + event.logIndex.toString())
